monitor/requirements.txt
requests==2.31.0
docker==6.1.3
python-dotenv==1.0.0